Version C0.36
RayRecord struct has been modified to store only a pointer for a ZeroSet so
that the necessary space for the set is allocated each time.  This saves a
space for storing each RayRecord ZeroSet.  For this modification, the setoper
library must have been changed so that set_initialize allocates the minimum
space.  Note that this new version (Jan. 23, 1994) does not work with the older
cdd programs.

Imported in Apr 2018 from cdd-036.tar.Z released Jan 23, 1994.

The following versions in between 0.33 and 0.36 had no released tarballs in Apr
2018:

Version C0.34 (Jan 22, 1994)
----------------------------
adjacency option has been added to output the adjacency list of output.

Version C0.35 (Jan 23, 1994)
-----------------------------
RayRecord struct has been modified to store only a pointer for a Ray vector so
that the necessary space for the vector is allocated each time.  This saves a
space for storing each RayRecord.