Version 0.60
The following changes are equivalent to ones that had been made for cdd+-074.
The default output file names have been changed to be consistent with the
transformation. To avoid confusion, *.ine file should be used only for a system
of linear inequalities, and *.ext file only for a set of extreme points and
rays. Accordingly, the files *.ead (previously *.adj) and *.ecd (previously
*.icd) are reserved for the adjacency and incidence files for the extremal
vertices/rays. Similarly, *.iad (previously *.iad) and *.icd (previously none)
are reserved for the adjacency and incidence files for the inequality data.

The LP code is now independent of cdd, and rewritten as a C library. This
library is called dplex, and contains two algorithms, the dual simplex and the
criss-cross method.

Imported in Apr 2018 from cdd-060.tar.gz released Aug 27, 1996.